Change filters on getting instance metadata to AND instead of OR
From the EC2 API on DescribeTags:

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeTags.html

You can specify multiple filters (for example, specify a specific
resource type and tag values that contain the string database). The
response includes information for a tag only if it matches all the
filters that you specified. If there's no match, no special message is
returned, the response is simply empty.

The phrase "only if it matches all the filters" means we should AND
search filters together.

Fixes bug #1190845

DocImpact

Change-Id: Icc0a7df1f93da803bbb2d7c9d18a4ce42bbe6720